requirements.txt
certifi==2022.9.24
influxdb-client==1.33.0
paho-mqtt==1.6.1
python-dateutil==2.8.2
reactivex==4.0.4
six==1.16.0
typing_extensions==4.3.0
urllib3==1.26.12
